requirements.txt
django=2.2.5
djangorestframework
psycopg2